Fixed the final reqs #7 , project is finished.
Woooooo!

+++ cancel_res.php
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>Canceling reservation...</title>
+        <title>Editing reservation...</title>
     </head>
     <body>
     <?php 
@@ -8,6 +8,7 @@
     	session_start();
    		$id = $_SESSION['activity_id'];
    		$userName = $_SESSION['username'];
+   		$res_count = $_SESSION['res_count'];
      	$query = "SELECT * FROM activity WHERE activity_id = '$id'";
      	$result = mysqli_query($conn,$query);
      	if (!$result) {
@@ -16,18 +17,45 @@
 			exit;
 		}
 		else{
-			$sqlUpdate = "UPDATE activity SET fullness = fullness - 1 WHERE activity_id = '$id'";
-			$sqlReserve = "DELETE FROM `reserve` WHERE user_name = '$userName' AND activity_id = '$id'";
-			$resultUpdate = mysqli_query($conn,$sqlUpdate);
-			$resultReserve = mysqli_query($conn,$sqlReserve);
-			if (!$resultUpdate || !$resultReserve) {
-		    	printf("\n\nError: %s\n", mysqli_error($conn));
-	    		exit();
+			if (isset($_POST['btnCancel'])) {
+				$sqlUpdate = "UPDATE activity SET fullness = fullness - '$res_count' WHERE activity_id = '$id'";
+				$sqlReserve = "DELETE FROM `reserve` WHERE user_name = '$userName' AND activity_id = '$id'";
+				$resultUpdate = mysqli_query($conn,$sqlUpdate);
+				$resultReserve = mysqli_query($conn,$sqlReserve);
+				if (!$resultUpdate || !$resultReserve) {
+			    	printf("\n\nError: %s\n", mysqli_error($conn));
+		    		exit();
+				}
+				else{
+					echo "Reservation has been cancelled!";
+					header('refresh:2; url=userPage.php');
+					exit;
+				}
 			}
 			else{
-				echo "Reservation has been cancelled!";
-				header('refresh:2; url=userPage.php');
-				exit;
+   				$res_count_update = $_POST['res_count_update'];
+				$sqlUpdate = "UPDATE activity SET fullness = fullness - ('$res_count' - '$res_count_update') WHERE activity_id = '$id'";
+				$sqlReserve = "UPDATE `reserve` SET res_count = '$res_count_update' WHERE user_name = '$userName' AND activity_id = '$id'";
+				$resultUpdate = mysqli_query($conn,$sqlUpdate);
+				$resultReserve = mysqli_query($conn,$sqlReserve);
+				if (!$resultUpdate || !$resultReserve) {
+			    	printf("\n\nError: %s\n", mysqli_error($conn));
+		    		exit();
+				}
+				else{
+					echo "Reservation has been updated!";
+					$_SESSION['res_count'] = $res_count_update;
+					if($res_count_update == 0){
+						$sqlReserve = "DELETE FROM `reserve` WHERE user_name = '$userName' AND activity_id = '$id'";
+						$resultReserve = mysqli_query($conn,$sqlReserve);
+						if(!$resultReserve){	
+					    	printf("\n\nError: %s\n", mysqli_error($conn));
+				    		exit();
+						}
+					}
+					header('refresh:2; url=userPage.php');
+					exit;
+				}
 			}
 		}
    		$conn->close();
